Separate task utils from task functions
